extract GUIManager as private library
